added hera tables-go reader

+++ pkg/utils/logging/v0/err_log_wrappers.go
@@ -3,6 +3,6 @@ package v0
 import "github.com/rs/zerolog/log"
 
 func (l *LibV0) ErrHandler(e error) error {
-	log.Error().Err(e).Int64("UNIX UTC Time: ", l.UnixTimeStampNow())
+	log.Error().Err(e).Int("UNIX UTC Time: ", l.UnixTimeStampNow())
 	return e
 }
